backend_service/requirements.txt
pyserial>=3.5
fastapi>=0.135.3,<1.0
uvicorn>=0.43.0,<1.0
httpx>=0.28.1
PyMySQL>=1.1,<2.0
